+++ package.json
@@ -22,7 +22,8 @@
       "types": "./index.d.ts"
     },
     "./utilities": {
-      "import": "./utilities/index.mjs"
+      "import": "./utilities/index.mjs",
+      "types": "./utilities/index.d.ts"
     }
   },
   "main": "index.js",
